added new model structure
