+++ listen_darwin.py
@@ -39,36 +39,8 @@ class MyListener(Listener):
 
         else:
             s = Schedule()
-            s.uid = m.uid
-            s.rid = m.rid
-
-        s.headcode = m.headcode
-        s.start_date = m.start_date
-        s.toc_code = m.toc_code
-        s.category = m.category
-        s.status = m.status
-        s.active = m.active
-        s.deleted = m.deleted
-        s.cancel_tiploc = m.cancel_reason_tiploc
-        s.cancel_code = m.cancel_reason_code
-        s.cancel_near = m.cancel_reason_near
-        s.save()
-
-        for o in m.all_points:
-            p = CallingPoint()
-            p.tiploc = o.tiploc
-            p.schedule = s
-            p.activity_codes = o.planned_activity_codes
-            p.cancelled = o.cancelled
-            p.false_tiploc = o.false_tiploc
-            p.route_delay = o.route_delay
-            p.working_arrival = o.raw_working_arrival_time
-            p.working_pass = o.raw_working_pass_time
-            p.working_departure = o.raw_working_departure_time
-            p.public_arrival = o.raw_public_arrival_time
-            p.public_departure = o.raw_public_departure_time
-            p.type = str(type(o))
-            p.save()
+        
+        s.fromPyXb(m, save=True, recursive=True)
 
     @db.transaction()
     def on_deactivated_message(self, message):
@@ -157,73 +129,8 @@ class MyListener(Listener):
     @db.transaction()
     def on_train_status_message(self, message):
         print("^^^^^^^^^^^^ NEW TRAIN STATUS MESSAGE!! ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^")
-
-        late_reason = message.late_reason
-        lr = None
-
-        if late_reason:
-            lr = LateReason()
-            lr.code = late_reason.code
-            lr.tiploc = late_reason.tiploc
-            lr.near = late_reason.near
-            lr.save()
-
         t = TrainStatus()
-        t.late_reason = lr
-        t.rid = message.rid
-        t.uid = message.uid
-        t.start = message.start_date
-        t.reverse_formation = message.reverse_formation
-        t.save()
-
-        for location in message.locations:
-            # Building forecasts
-            f_arrival = None
-            f_departure = None
-            f_pass = None
-
-            if location.forecast_arrival_time:
-                f_arrival = build_forecast(location.forecast_arrival_time)
-                f_arrival.save()
-
-            if location.forecast_departure_time:
-                f_departure = build_forecast(location.forecast_departure_time)
-                f_departure.save()
-
-            if location.forecast_pass_time:
-                f_pass = build_forecast(location.forecast_pass_time)
-                f_pass.save()
-
-            # Building platform
-            platform = location.platform
-            p = None
-
-            if platform:
-                p = Platform()
-                p.source = platform.source
-                p.suppressed = platform.suppressed
-                p.suppressed_by_cis = platform.suppressed_by_cis
-                p.confirmed = platform.confirmed
-                p.number = platform.number
-                p.save()
-
-            # Building location
-            l = Location()
-            l.train_status = t
-            l.platform = p
-            l.forecast_arrival = f_arrival
-            l.forecast_departure = f_departure
-            l.forecast_pass = f_pass
-            l.tiploc = location.tiploc
-            l.suppressed = location.suppressed
-            l.detach_front = location.detach_front
-            l.working_arrival = location.working_arrival_time
-            l.working_departure = location.working_departure_time
-            l.working_pass = location.working_pass_time
-            l.public_arrival = location.public_arrival_time
-            l.public_departure = location.public_departure_time
-            l.length = location.length
-            l.save()
+        t.fromPyXb(message, save=True, recursive=True)
 
 def build_assoc_svc(assoc_svc):
     a = AssociationService()
@@ -260,14 +167,10 @@ def build_train_order_item(train_order):
 
 # Make this file importable for debugging
 
-class HPClient(Client):
-    def on_disconnected(self):
-        self.reconnect(retries=0, delay=10) # try to reconnect forever
-
 if __name__ == "__main__":
 
     # Instantiate the Push Port client.
-    client = HPClient(os.environ["STOMP_USER"],
+    client = Client(os.environ["STOMP_USER"],
                     os.environ["STOMP_PASS"],
                     os.environ["STOMP_QUEUE"],
                     MyListener)
