+++ models/DeactivatedSchedule.py
@@ -1,0 +1,5 @@
+from peewee import CharField
+from models import DarwinModel
+
+class DeactivatedSchedule(DarwinModel):
+    rid                 = CharField()
